+++ __brick__/lib/features/{{feature_name}}/domain/data_sources/{{feature_name}}_data_source_local.dart
@@ -1,8 +1,11 @@
 import '../../usecases/{{usecase_action}}_{{entity_name}}_usecase.dart';
-import '../interfaces/{{feature_name}}_data_source_interface.dart';
 import '../entities/{{entity_name}}_class.dart';
+import '../interfaces/{{feature_name}}_data_source_interface.dart';
+import '../../../../core/interfaces/base_data_source.dart';
 
-class {{#pascalCase}}{{feature_name}}{{/pascalCase}}DataSourceLocal implements {{#pascalCase}}{{feature_name}}{{/pascalCase}}DataSourceInterface {
+class {{#pascalCase}}{{feature_name}}{{/pascalCase}}DataSourceLocal extends LocalDataSource implements {{#pascalCase}}{{feature_name}}{{/pascalCase}}DataSourceInterface {
+  {{#pascalCase}}{{feature_name}}{{/pascalCase}}DataSourceLocal(super.keyValueStore);
+
   @override
   Future<{{#pascalCase}}{{entity_name}}{{/pascalCase}}?> {{#camelCase}}{{usecase_action}}{{/camelCase}}{{#pascalCase}}{{entity_name}}{{/pascalCase}}(
     {{#pascalCase}}{{usecase_action}}{{/pascalCase}}{{#pascalCase}}{{entity_name}}{{/pascalCase}}Request request,