+++ __brick__/lib/features/{{feature_name}}/domain/data_sources/{{feature_name}}_data_source_remote.dart
@@ -1,5 +1,6 @@
+import 'package:{{project_name}}/features/{{feature_name}}/usecases/{{usecase_action}}_{{entity_name}}_usecase.dart';
+
 import '../../../../core/interfaces/base_data_source.dart';
-import '../../usecases/{{usecase_action}}_{{entity_name}}_usecase.dart';
 import '../entities/{{entity_name}}_class.dart';
 import '../interfaces/{{feature_name}}_data_source_interface.dart';
 
@@ -10,7 +11,7 @@ class {{#pascalCase}}{{feature_name}}{{/pascalCase}}DataSourceRemote extends Rem
   Future<{{#pascalCase}}{{entity_name}}{{/pascalCase}}?> {{#camelCase}}{{usecase_action}}{{/camelCase}}{{#pascalCase}}{{entity_name}}{{/pascalCase}}(
     {{#pascalCase}}{{usecase_action}}{{/pascalCase}}{{#pascalCase}}{{entity_name}}{{/pascalCase}}Request request,
   ) async {
-    // TODO: set the correct API path and verb (get/post).
+    // TODO: set the correct API path and verb (get/post/patch/delete).
     final nr = await apiService.get(
       '{{feature_name}}',
       enableLogs: true,